+++ source/bot.js
@@ -92,7 +92,7 @@ dAPIClient.login(require(authPath).token);
 
 //#region Event Handlers
 dAPIClient.on(Events.ClientReady, () => {
-	console.log(`Connected as ${dAPIClient.user.tag}`);
+	console.log(`Connected as ${dAPIClient.user.tag} -- Run Mode: ${runMode}`);
 	if (runMode === "production") {
 		(() => {
 			try {
@@ -147,6 +147,12 @@ dAPIClient.on(Events.ClientReady, () => {
 
 dAPIClient.on(Events.InteractionCreate, async interaction => {
 	await dbReady;
+	const [interactingHunter] = await logicLayer.hunters.findOrCreateBountyHunter(interaction.user.id, interaction.guild.id);
+	if (interactingHunter.isBanned && !(interaction.isCommand() && interaction.commandName === "moderation")) {
+		interaction.reply({ content: `You are banned from interacting with BountyBot on ${interaction.guild.name}.`, flags: [MessageFlags.Ephemeral] });
+		return;
+	}
+
 	if (interaction.isAutocomplete()) {
 		const command = getCommand(interaction.commandName);
 		const focusedOption = interaction.options.getFocused(true);
@@ -253,9 +259,9 @@ dAPIClient.on(Events.GuildDelete, async guild => {
 	logicBlob.goals.deleteCompanyGoals(guild.id);
 	logicBlob.hunters.deleteCompanyHunters(guild.id);
 	(await logicBlob.ranks.findAllRanks(guild.id))
-			.map(r => r.roleId)
-			.filter(id => !!id)
-			.forEach(id => guild.roles.delete(id, 'Cleaning up BountyBot roles during kick.'));
+		.map(r => r.roleId)
+		.filter(id => !!id)
+		.forEach(id => guild.roles.delete(id, 'Cleaning up BountyBot roles during kick.'));
 	logicBlob.ranks.deleteCompanyRanks(guild.id);
 	logicBlob.companies.deleteCompany(guild.id);
 });